Adding new widget placeholder component.
